python_backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
pandas>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10
openpyxl>=3.1.0  